requirements.txt
uvicorn
fastapi
httpx
pyyaml